refactor(stage7-BC): net::room 切到 GameEngine + AI 删 legacy 桥
Phase B + C 合并 (耦合紧):
- src/ai/dummy.rs: 删 ai_choose_discard_legacy / ai_react_to_discard_legacy
- src/ai/timeout.rs: 删 default_action_on_timeout_legacy
- src/net/room.rs: RoomActor.game 字段 GameState → GameEngine; 调用站点
  全部走 method 形式 (turn() / phase() / players() / last_discard()
  / round_wind() / dealer() / kyoku() / honba() / riichi_sticks()
  / wall_remaining() / dora_indicators()). do_*/declare_*/try_* 系列
  方法名签名一致, 大多机械替换. apply_ai_action 用 ai_choose_discard
  替代 legacy.
- src/net/protocol.rs: GameStateView.riichi_sticks 类型 u8 → u32 (跟随
  MatchState.riichi_sticks_pool 类型).
- src/net/room.rs::tests mod 用 #[cfg(all(test, feature = "net-tests"))]
  屏蔽 — 测试大量直接戳 GameState 内部字段, GameEngine 不允许这种黑魔法,
  重写为 round_apply 驱动留 follow-up.
- Cargo.toml 加 feature net-tests (默认关).

GameStateView 仍保留为网络协议数据格式 (与 GameState 不绑定, 仅命名沿用).

cargo build --bin tui-majo + cargo test --lib --no-run 通过.

+++ src/net/protocol.rs
@@ -178,7 +178,7 @@ pub struct GameStateView {
     pub round_wind: RoundWind,
     pub kyoku: u8,
     pub honba: u8,
-    pub riichi_sticks: u8,
+    pub riichi_sticks: u32,
     pub dealer: Seat,
     pub turn: Seat,
     pub phase: Phase,
